Clean P2P unit tests
