+++ elements/lisk-p2p/test/unit/exceptions.ts
@@ -12,7 +12,6 @@
  * Removal or modification of this copyright notice is prohibited.
  *
  */
-
 import { expect } from 'chai';
 import {
 	InvalidPeerException,
